Enable -Werror

+++ projects/rocshmem/src/util.hpp
@@ -180,17 +180,17 @@ typedef struct device_prop {
 extern std::vector<device_prop_t> device_properties;
 
 [[maybe_unused]] static int get_threads_per_block(int device_id) {
-  assert(device_properties.size() > device_id);
+  assert(static_cast<int>(device_properties.size()) > device_id);
   return device_properties[device_id].maxThreadsPerBlock;
 }
 
 [[maybe_unused]] static int get_wf_size(int device_id) {
-  assert(device_properties.size() > device_id);
+  assert(static_cast<int>(device_properties.size()) > device_id);
   return device_properties[device_id].warpSize;
 }
 
 [[maybe_unused]] static const char* get_arch_name(int device_id) {
-  assert(device_properties.size() > device_id);
+  assert(static_cast<int>(device_properties.size()) > device_id);
   return device_properties[device_id].gcnArchName;
 }
 
@@ -445,7 +445,7 @@ template <typename... Args>
   uint8_t* dst_bytes{static_cast<uint8_t*>(dst)};
   uint8_t* src_bytes{static_cast<uint8_t*>(src)};
 
-  for (int i = 16; i > 1; i >>= 1) {
+  for (size_t i = 16; i > 1; i >>= 1) {
     while (size >= i) {
       store_asm(src_bytes, dst_bytes, i);
       src_bytes += i;
